Split StardustClientExt into StardustIdentityClient/Base traits
