+++ identity_stardust/examples/ex2_resolve_did.rs
@@ -1,10 +1,11 @@
 // Copyright 2020-2022 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-use identity_stardust::StardustClientExt;
-use identity_stardust::StardustDocument;
 use iota_client::block::output::AliasOutput;
 use iota_client::Client;
+
+use identity_stardust::StardustDocument;
+use identity_stardust::StardustIdentityClient;
 
 mod ex0_create_did;
