+++ identity_stardust/src/client/mod.rs
@@ -1,6 +1,12 @@
 // Copyright 2020-2022 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-pub use client_ext::StardustClientExt;
+pub use identity_client::StardustIdentityClient;
+pub use identity_client::StardustIdentityClientBase;
 
-mod client_ext;
+#[cfg(feature = "iota-client")]
+pub use self::iota_client::StardustClientExt;
+
+mod identity_client;
+#[cfg(feature = "iota-client")]
+mod iota_client;
