+++ identity_stardust/src/lib.rs
@@ -4,7 +4,7 @@
 #![forbid(unsafe_code)]
 #![allow(clippy::upper_case_acronyms)]
 
-#[cfg(feature = "iota-client")]
+#[cfg(feature = "client")]
 pub use client::*;
 pub use did::StardustDID;
 pub use did::StardustDIDUrl;
@@ -15,7 +15,7 @@ pub use state_metadata::*;
 pub use self::error::Error;
 pub use self::error::Result;
 
-#[cfg(feature = "iota-client")]
+#[cfg(feature = "client")]
 mod client;
 mod did;
 mod document;
